Refactor event and query for number selection
Previously the numbers were inputted in the calculator by going through
the event and getting the content inside the specific target in the
event.

It now has changed to simply manage numbers (as it just did that
before), by directly assigning the event listener to the number itself,
instead of searching in the target of the event.